django3-project/demosite/requirements.txt
# Django
Django==3.1.14

# Rest API
djangorestframework==3.12.2

# Set CORS headers
django-cors-headers==3.6.0

# Production deployment
gunicorn==20.0.4
whitenoise==5.2.0
